Ajust Register Pages
