+++ Front-end/package.json
@@ -12,6 +12,7 @@
     "@unform/core": "^2.1.6",
     "@unform/web": "^2.1.6",
     "axios": "^0.21.4",
+    "input-mask": "^2.0.0",
     "next": "11.0.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
