+++ Front-end/src/components/Form/inputMask.tsx
@@ -1,0 +1,7 @@
+
+
+const InputMask = () => {
+
+}
+
+export default InputMask;
