+++ Front-end/src/components/Register/FormRegister/index.tsx
@@ -1,0 +1,173 @@
+import { NextPage } from "next";
+import { Form } from '@unform/web';
+import { Scope } from '@unform/core';
+
+import Input from '../../Form/input';
+
+import { toast } from 'react-toastify';
+
+import {
+  HeaderTitle,
+  RowInputs,
+  InputsBox,
+} from './style';
+
+import { useRef } from "react";
+
+
+const FormRegister: NextPage = () => {
+
+  const genreRef = useRef<HTMLSelectElement>(null);
+
+  const handleSubmit = (data) => {
+    /* const genre = genreRef.current?.value;
+
+    let idUser;
+
+    const dataUser = {
+      ...data.user,
+      genre,
+    };
+
+    api
+      .post('/users', dataUser)
+      .then((res) => {
+        idUser = res.data.id;
+
+        const dataAdresses = {
+          idUser,
+          ...data.address,
+        };
+
+        api
+          .post('/addresses', dataAdresses)
+          .then((res) => {
+            toast.success('Usuario cadastrado com Sucesso!');
+
+            setTimeout(() => {
+              router.push('/');
+            }, 2000);
+          })
+          .catch((err) => {
+            console.log(err.message);
+            return toast.error('Erro ao cadastrar o Usuario');
+          });
+      })
+      .catch((err) => {
+        console.log(err.message);
+        return toast.error('Usuario já existente!');
+      }); */
+  }
+
+  return (
+    <>
+      <HeaderTitle>Registre-se</HeaderTitle>
+      <Form onSubmit={handleSubmit}>
+        <Scope path="user">
+          <RowInputs>
+            <InputsBox fixedSize={'50%'} sizeResponsive={'100%'}>
+              <span>Nome</span>
+              <Input name="name" type="text" placeholder="Nome Completo" />
+            </InputsBox>
+            <InputsBox fixedSize={'25%'} sizeResponsive={'40%'}>
+              <span>Gênero</span>
+              <select ref={genreRef}>
+                <option value="Masculino">Masculino</option>
+                <option value="Feminino">Feminino</option>
+              </select>
+            </InputsBox>
+            <InputsBox fixedSize={'25%'} sizeResponsive={'55%'}>
+              <span>Data de Nascimento</span>
+              <Input
+                name="birthDate"
+                type="date"
+                placeholder="Data de Nascimento"
+              />
+            </InputsBox>
+          </RowInputs>
+
+          <RowInputs>
+            <InputsBox fixedSize={'50%'} sizeResponsive={'40%'}>
+              <span>RG</span>
+              <Input
+                name="rg"
+                type="text"
+                placeholder="_ _-_ _ _-_ _ _-_"
+              />
+            </InputsBox>
+            <InputsBox fixedSize={'50%'} sizeResponsive={'55%'}>
+              <span>CPF</span>
+              <Input
+                name="cpf"
+                type="text"
+                placeholder="_ _ _-_ _ _-_ _ _-_ _"
+              />
+            </InputsBox>
+          </RowInputs>
+
+          <RowInputs>
+            <InputsBox fixedSize={'50%'} sizeResponsive={'40%'}>
+              <span>Telefone</span>
+              <input
+                name="telephone"
+                type="text"
+                placeholder="(_ _)_ _ _ _-_ _ _ _"
+              />
+            </InputsBox>
+            <InputsBox fixedSize={'50%'} sizeResponsive={'55%'}>
+              <span>E-mail</span>
+              <Input
+                name="email"
+                type="text"
+                placeholder="contato@purple.com.br"
+              />
+            </InputsBox>
+            {/* <InputsBox fixedSize={'30%'} sizeResponsive={'40%'}>
+                  <span>Senha</span>
+                  <Input name="password" type="password" placeholder="Senha" />
+                </InputsBox> */}
+          </RowInputs>
+        </Scope>
+        <Scope path="address">
+          <RowInputs>
+            <InputsBox fixedSize={'50%'} sizeResponsive={'100%'}>
+              <span>CEP</span>
+              <Input name="cep" type="text" placeholder="_ _ _ _ _-_ _ _" />
+            </InputsBox>
+            <InputsBox fixedSize={'10%'} sizeResponsive={'45%'}>
+              <span>UF</span>
+              <Input name="district" type="text" placeholder="SP" />
+            </InputsBox>
+            <InputsBox fixedSize={'40%'} sizeResponsive={'50%'}>
+              <span>Cidade</span>
+              <Input name="city" type="text" placeholder="São Paulo" />
+            </InputsBox>
+          </RowInputs>
+
+          <RowInputs>
+            <InputsBox fixedSize={'50%'} sizeResponsive={'45%'}>
+              <span>Rua/Logradouro</span>
+              <Input
+                name="place"
+                type="text"
+                placeholder="Rua das Laranjeiras"
+              />
+            </InputsBox>
+            <InputsBox fixedSize={'10%'} sizeResponsive={'15%'}>
+              <span>Número</span>
+              <Input name="number" type="text" placeholder="290" />
+            </InputsBox>
+            <InputsBox fixedSize={'40%'} sizeResponsive={'30%'}>
+              <span>Complemento</span>
+              <Input name="complement" type="text" placeholder="Casa" />
+            </InputsBox>
+          </RowInputs>
+        </Scope>
+
+        <button>Continuar</button>
+      </Form>
+    </>
+  )
+}
+
+export default FormRegister;
